Refactor: extract sparse Zippel functions from modulargcd.py to zippel.py
